integrated API test and search for defendent from UI
