+++ testdata/cp/validjson/onehearning.json
@@ -1,0 +1,147 @@
+{
+  "hearing": {
+    "id": "{0}",
+    "jurisdictionType": "MAGISTRATES",
+    "courtCentre": {
+      "address": {
+        "address1": "redacted",
+        "address2": "redacted",
+        "address3": "redacted",
+        "address4": "",
+        "address5": "",
+        "postcode": "redacted"
+      },
+      "code": "B43KB00",
+      "courtHearingLocation": "B43KB03",
+      "id": "9bafd21f-71cc-3a12-807e-b8dca555cb71",
+      "lja": {
+        "ljaCode": "1922",
+        "ljaName": "redacted Magistrates' Court"
+      },
+      "name": "redacted Magistrates' Court",
+      "roomId": "e329e8c9-02f8-34e5-a77a-65bcdb760622",
+      "roomName": "Courtroom 03"
+    },
+    "type": {
+      "description": "First hearing",
+      "id": "4a0e892d-c0c5-3c51-95b8-704d8c781776"
+    },
+    "hearingDays": [
+      {
+        "courtCentreId": "9bafd21f-71cc-3a12-807e-b8dca555cb71",
+        "courtRoomId": "e329e8c9-02f8-34e5-a77a-65bcdb760622",
+        "isCancelled": false,
+        "listedDurationMinutes": 20,
+        "listingSequence": 6,
+        "sittingDay": "2025-01-09T10:00:00.000Z"
+      }
+    ],
+    "hearingLanguage": "ENGLISH",
+    "prosecutionCases": [
+      {
+        "id": "2ff59fb9-c61d-4e1b-a225-3d6a79b2088a",
+        "prosecutionCaseIdentifier": {
+          "address": {
+            "address1": "redacted",
+            "address2": "redacted",
+            "address3": "redacted",
+            "address4": "redacted",
+            "address5": "redacted",
+            "postcode": "redacted"
+          },
+          "contact": {
+            "primaryEmail": "redacted@redacted.co.uk"
+          },
+          "majorCreditorCode": "redacted",
+          "prosecutionAuthorityCode": "redacted",
+          "prosecutionAuthorityId": "e86becd5-aae8-3626-9789-188666240a39",
+          "prosecutionAuthorityName": "redacted Police",
+          "prosecutionAuthorityOUCode": "0430000",
+          "caseURN": "redacted"
+        },
+        "initiationCode": "Q",
+        "defendants": [
+          {
+            "id": "5e8fcf7e-6884-4789-b6ff-be63ad149754",
+            "prosecutionCaseId": "2ff59fb9-c61d-4e1b-a225-3d6a79b2088a",
+            "offences": [
+              {
+                "id": "6cd83056-6ba9-4007-b796-c4b5dadcbfbf",
+                "offenceDefinitionId": "fdab1c88-5fe2-3cb3-962b-1a1e79d976b9",
+                "offenceCode": "redacted",
+                "offenceTitle": "Assault by beating",
+                "wording": "Assault by beating (recordable), redacted, redacted, 105/01 / \n\nOn 19/06/2024 at redacted in the County of redacted assaulted redacted redacted by beating redacted.\n\n ",
+                "startDate": "2024-06-19",
+                "offenceTitleWelsh": "Adran 39 - ymosod drwy guro (curfa)",
+                "offenceLegislation": "Contrary to section 39 of the Criminal Justice Act1988.",
+                "offenceLegislationWelsh": "Yn groes i adran 39 Deddf Cyfiawnder Troseddol 1988.",
+                "modeOfTrial": "Summary",
+                "arrestDate": "2024-06-20",
+                "chargeDate": "2024-11-23",
+                "orderIndex": 1,
+                "count": 0,
+                "listingNumber": 1
+              }
+            ],
+            "prosecutionAuthorityReference": "redacted",
+            "personDefendant": {
+              "arrestSummoredactedNumber": "redacted",
+              "personDetails": {
+                "address": {
+                  "address1": "redacted",
+                  "address2": "redacted",
+                  "address3": "redacted",
+                  "address4": "redacted",
+                  "address5": "redacted",
+                  "postcode": "redacted"
+                },
+                "contact": {
+                  "mobile": "redacted"
+                },
+                "dateOfBirth": "1990-01-01",
+                "documentationLanguageNeeds": "ENGLISH",
+                "ethnicity": {
+                  "observedEthnicityCode": "",
+                  "observedEthnicityDescription": "redacted",
+                  "observedEthnicityId": "c4ca4238-a0b9-3382-8dcc-509a6f75849b",
+                  "selfDefinedEthnicityCode": "redacted",
+                  "selfDefinedEthnicityDescription": "Not stated",
+                  "selfDefinedEthnicityId": "b8d13cd7-99b1-8888-a3ca-b8974ca9702f"
+                },
+                "firstName": "xyz",
+                "gender": "MALE",
+                "hearingLanguageNeeds": "ENGLISH",
+                "lastName": "redacted"
+              }
+            },
+            "aliases": [
+              {
+                "firstName": "redacted",
+                "lastName": "redacted"
+              }
+            ],
+            "pncId": "redacted",
+            "masterDefendantId": "5e8fcf7e-6884-4789-b6ff-be63ad149754",
+            "isYouth": false
+          }
+        ],
+        "caseStatus": "ACTIVE",
+        "originatingOrganisation": "redacted"
+      }
+    ],
+    "hasSharedResults": false,
+    "judiciary": [
+      {
+        "firstName": "redacted",
+        "judicialId": "b595f793-d103-34b1-9b5f-7497c85773ed",
+        "judicialRoleType": {
+          "judicialRoleTypeId": "b595f793-d103-34b1-9b5f-7497c85773ed",
+          "judiciaryType": "District Judge (MC)"
+        },
+        "lastName": "redacted",
+        "title": "",
+        "userId": "67964394-8dd5-419c-b2d1-64c1bb9db24f"
+      }
+    ]
+  }
+}
